+++ src/main/java/com/vansl/controller/BlogController.java
@@ -1,9 +1,11 @@
 package com.vansl.controller;
 
-import com.alibaba.fastjson.JSONObject;
 import com.vansl.dto.TableData;
 import com.vansl.entity.Blog;
 import com.vansl.service.BlogService;
+import com.vansl.service.UserService;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,16 +28,21 @@ public class BlogController {
     @Autowired
     BlogService blogService;
 
+    @Autowired
+    UserService userService;
+
     // 查询用户的所有文章（包括草稿）
     @GetMapping(params = {"userId","offset","limit"})
     @ResponseBody
     public TableData getAllData(Integer userId, Integer offset, Integer limit){
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=userId){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        if(userId!=loginId){
+            return null;
+        }
+
         return  blogService.selectAll(userId,false,offset,limit);
     }
 
@@ -43,12 +50,16 @@ public class BlogController {
     @GetMapping("/{id}")
     @ResponseBody
     public String getContent(@PathVariable Integer id){
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=userId){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        Integer userId=blogService.selectUserIdByBlogId(id);
+        if(userId!=loginId){
+            System.out.println(userId+" "+loginId);
+            return "denied";
+        }
+
         return  blogService.selectContentByBlogId(id);
     }
 
@@ -56,13 +67,12 @@ public class BlogController {
     @PostMapping
     @ResponseBody
     public String addBlog(@RequestBody Blog blog, HttpServletResponse response ){
-        Integer userId=blog.getUserId();
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=userId){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        blog.setUserId(loginId);
+
         Integer result=blogService.insertBlog(blog);
         //操作失败返回则错误信息
         if (result==-1){
@@ -77,13 +87,15 @@ public class BlogController {
     @PutMapping("/{id}")
     @ResponseBody
     public String updateBlog(@PathVariable Integer id,@RequestBody Blog blog, HttpServletResponse response ){
-        Integer userId=blog.getUserId();
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=userId){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        Integer userId=blogService.selectUserIdByBlogId(id);
+        if(userId!=loginId){
+            return "denied";
+        }
+
         blog.setId(id);
         Integer result=blogService.updateBlog(blog);
         //操作失败返回则错误信息
@@ -98,16 +110,15 @@ public class BlogController {
     // 删除博客
     @DeleteMapping("/{id}")
     @ResponseBody
-    public String deleteBlogComment(@PathVariable("id")Integer id, @RequestBody String data, HttpServletResponse response ){
-        // 把请求数据转换成请求对象
-        JSONObject json=JSONObject.parseObject(data);
-        Integer userId=(Integer) json.get("userId");
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=userId){
-         *      return "denied";
-         *  }
-         */
+    public String deleteBlog(@PathVariable("id")Integer id, HttpServletResponse response ){
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        Integer userId=blogService.selectUserIdByBlogId(id);
+        if(userId!=loginId){
+            return "denied";
+        }
 
         Integer result=blogService.deleteBlog(id);
         //操作失败返回则错误信息
