+++ src/main/java/com/vansl/controller/BlogTypeController.java
@@ -1,10 +1,11 @@
 package com.vansl.controller;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONObject;
 import com.vansl.dto.TypeTreeNode;
 import com.vansl.entity.BlogType;
 import com.vansl.service.BlogTypeService;
+import com.vansl.service.UserService;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -23,6 +24,9 @@ public class BlogTypeController {
     @Autowired
     BlogTypeService blogTypeService;
 
+    @Autowired
+    UserService userService;
+
     // 通过id查询博客分类
     @GetMapping("/{id}")
     @ResponseBody
@@ -31,10 +35,15 @@ public class BlogTypeController {
     }
 
     // 查询用户的所有分类数据
-    @GetMapping(params = {"userId"})
+    @GetMapping
     @ResponseBody
-    public List<TypeTreeNode> getAllData(@RequestParam Integer userId){
-        return blogTypeService.selectAll(userId);
+    public List<TypeTreeNode> getAllData(){
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+
+        return blogTypeService.selectAll(loginId);
     }
 
     // 通过博客id查询博客分类
@@ -49,12 +58,11 @@ public class BlogTypeController {
     @PostMapping
     @ResponseBody
     public String addBlogType(@RequestBody BlogType type, HttpServletResponse response){
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=type.getUserId()){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        type.setUserId(loginId);
 
         //调用service执行操作
         Integer result=blogTypeService.insertBlogType(type);
@@ -74,12 +82,14 @@ public class BlogTypeController {
     @PutMapping
     @ResponseBody
     public String updateBlogType(@RequestBody BlogType type, HttpServletResponse response){
-        /*登录验证功能尚未实现
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=type.getUserId()){
-         *      return "denied";
-         *  }
-         */
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        Integer userId=blogTypeService.selectUserIdByTypeId(type.getId());
+        if(userId!=loginId){
+            return "denied";
+        }
 
         //调用service执行操作
         Integer result=blogTypeService.updateBlogType(type);
@@ -95,14 +105,20 @@ public class BlogTypeController {
     // 删除博客分类
     @DeleteMapping("/{id}")
     @ResponseBody
-    public String deleteBlogType(@PathVariable("id")Integer id,@RequestBody BlogType type, HttpServletResponse response){
-        /*登录验证功能尚未实现*
-         *  Integer loginId=redisUtil.getUser().getUserId();
-         *  if(loginId!=type.getUserId()){
-         *      return "denied";
-         *  }
-         */
+    public String deleteBlogType(@PathVariable("id")Integer id, HttpServletResponse response){
+        //身份验证
+        Subject subject = SecurityUtils.getSubject();
+        String userName=(String)subject.getPrincipals().getPrimaryPrincipal();
+        Integer loginId=userService.selectIdByUserName(userName);
+        Integer userId=blogTypeService.selectUserIdByTypeId(id);
+        if(userId!=loginId){
+            return "denied";
+        }
+
+        BlogType type=new BlogType();
         type.setId(id);
+        type.setUserId(userId);
+
         //调用service执行操作
         Integer result=blogTypeService.deleteBlogType(type);
         //操作失败返回则错误信息
